+++ railties/lib/rails/application/bootstrap.rb
@@ -53,7 +53,11 @@ INFO
           logger
         end
 
-        Rails.logger.level = ActiveSupport::Logger.const_get(config.log_level.to_s.upcase)
+        if ::Logger === Rails.logger
+          Rails.logger.level = ActiveSupport::Logger.const_get(config.log_level.to_s.upcase)
+        else
+          Rails.logger.level = config.log_level
+        end
       end
 
       # Initialize cache early in the stack so railties can make use of it.